Update to work with Travis-CI container builds
This removes the need for sudo and lets builds start quicker
as they don't have to spin up a whole virtual machine. The
install of virtualenv is moved into the apt-get section of
.travis.yml and before_install.sh now only installs tools
if they aren't cached.

Tested with my resistor-time repo, verified to work on
first build and when tools are reloaded from cache.